WIP workspaces for core & types packages
