+++ package.json
@@ -15,7 +15,7 @@
     "test": "npm run -s test:code && npm run -s test:typescript && npm run -s test:coverage",
     "test:code": "c8 node test.js",
     "test:coverage": "c8 check-coverage",
-    "test:typescript": "tsd && for d in tests/* ; do tsd $d; done"
+    "test:typescript": "for d in tests/* ; do tsd $d; done"
   },
   "renovate": {
     "extends": [
@@ -33,8 +33,10 @@
     "uvu": "^0.5.1"
   },
   "workspaces": [
+    "packages/types",
+    "packages/core",
     "packages/types-rest-api-github.com",
-    "packages/types-rest-api-3.1",
-    "packages/types-rest-api-3.0"
+    "packages/types-rest-api-ghes-3.1",
+    "packages/types-rest-api-ghes-3.0"
   ]
 }
