+++ tests/ghes-3.0/index.js
@@ -1,4 +1,4 @@
-import { Octokit } from "../../index.js";
+import { Octokit } from "@octokit-next/core";
 
 export const OctokitGhes30 = Octokit.withDefaults({
   version: "ghes-3.0",
