+++ tsconfig.json
@@ -4,5 +4,5 @@
     "strict": true,
     "strictNullChecks": true
   },
-  "exclude": ["**"]
+  "include": []
 }
